Adding updated version of front-end with elastic beanstalk v0.1
