+++ .ebextensions/nodecommand.config
@@ -1,0 +1,5 @@
+// .ebextensions/nodecommand.config
+//
+//option_settings:
+//  aws:elasticbeanstalk:container:nodejs:
+//    NodeCommand: "node server.compiled.js"
