+++ .ebextensions/proxy.config
@@ -1,0 +1,3 @@
+location /static {
+  alias /var/app/current/client/build/static;
+}
