+++ config/config.json
@@ -6,11 +6,12 @@
     "host": "127.0.0.1",
     "dialect": "mysql"
   },
-  "test": {
-    "username": "root",
-    "password": null,
-    "database": "database_test",
-    "host": "127.0.0.1",
+  "staging": {
+    "username": "staging",
+    "password": "cascarita2023!",
+    "database": "test_db",
+    "host": "awseb-e-ismrimxiv2-stack-awsebrdsdatabase-3kjgpageucrn.ctxji11ur79e.us-east-2.rds.amazonaws.com",
+    "port": 3306,
     "dialect": "mysql"
   },
   "production": {
