+++ package.json
@@ -1,15 +1,10 @@
 {
-  "name": "nodemvc",
-  "version": "1.0.0",
-  "description": "",
-  "main": "index.js",
+  "name": "cascarita",
+  "version": "0.0.1",
+  "private": true,
   "scripts": {
-    "start": "nodemon server.js --ignore client",
-    "test": "echo \"Error: no test specified\" && exit 1"
+    "start": "node app"
   },
-  "keywords": [],
-  "author": "",
-  "license": "ISC",
   "dependencies": {
     "express": "^4.18.2",
     "mysql2": "^3.6.5",
